Message delivery works PERFECTLY. Concurrency model with overwriting protection added, timeDelivered is now concurrent and happens automatically. Added diferent broadcast message types instead of 1

+++ src/net/recv.rs
@@ -22,7 +22,7 @@ pub enum ClientMessageIn {
     Logout,
 
     /// A list of MessageId's that the client reports to see
-    SeeMessages(Vec<u128>),
+    SeeMessages(Vec<u32>),
     SendMessage(ChatMessageSender),
 
     /// Client can send this to server to fetch old messages.
